codes/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# AI/ML Libraries
openai==1.3.7
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
nltk==3.8.1

# Data Validation
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Security (optional, for future auth features)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
